finish register user function

+++ db.sql
@@ -1,7 +1,7 @@
 CREATE DATABASE assignment4;
 
 CREATE TABLE Users (
-    id integer PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     username text,
     email text,
     password text,
@@ -9,7 +9,7 @@ CREATE TABLE Users (
 );
 
 CREATE TABLE Photos (
-    id numeric PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     title text,
     caption VARCHAR(50),
     photo_url VARCHAR(50),
@@ -17,14 +17,14 @@ CREATE TABLE Photos (
 );
 
 CREATE TABLE Comments (
-    id numeric PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     user_id integer REFERENCES Users (id),
     photo_id integer REFERENCES Photos (id),
     message VARCHAR(255)
 );
 
 CREATE TABLE SocialMedias (
-    id numeric PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     user_id integer REFERENCES Users (id),
     name text,
     social_media_url text
